requirements.txt
alembic==1.4.3
click==7.1.2
ecdsa==0.14.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jose==3.2.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
